move SearchIndexSynchronousProtocol to its own test file
